+++ lib/libft.h
@@ -6,7 +6,7 @@
 /*   By: dkremer <dkremer@student.42heilbronn.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 15:28:54 by dkremer           #+#    #+#             */
-/*   Updated: 2023/10/19 23:17:54 by dkremer          ###   ########.fr       */
+/*   Updated: 2023/10/19 23:47:49 by dkremer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int				ft_isalpha(int c);
 int				ft_isascii(int a);
 int				ft_isdigit(int i);
 int				ft_isprint(int p);
-int				ft_strlen(const char *str);
+size_t			ft_strlen(const char *str);
 int				ft_strncmp(const char *s1, const char *s2, size_t n);
 int				ft_tolower(int c);
 int				ft_toupper(int c);
@@ -30,10 +30,10 @@ char			*ft_strchr(const char *s, int c);
 char			*ft_strrchr(const char *s, int c);
 void			*ft_memset(void *ptr, int value, size_t num);
 void			ft_bzero(void *s, unsigned int n);
-void			*ft_memcpy(void *dst, const void *src, unsigned int n);
+void			*ft_memcpy(void *dst, const void *src, size_t n);
 void			*ft_memmove(void *dst, const void *src, unsigned int len);
-unsigned int	ft_strlcpy(char *dst, const char *src, size_t dstsize);
-unsigned int	ft_strlcat(char *dst, const char *src, unsigned int dstsize);
+size_t			ft_strlcpy(char *dst, const char *src, size_t dstsize);
+size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
 void			*ft_memchr(const void *s, int c, unsigned int n);
 int				ft_memcmp(const void *s1, const void *s2, unsigned int n);
 char			*ft_strnstr(const char *hay, const char *nee, size_t len);
